振動・音声機能のファイルを分けた
