+++ LogiQuiz/Images.swift
@@ -1,0 +1,13 @@
+//
+//  Images.swift
+//  LogiQuiz
+//
+//  Created by KentoFujita on 2023/07/20.
+//
+
+import UIKit
+
+enum Images {
+    static let correct = UIImage(named: "correct")
+    static let incorrect = UIImage(named: "incorrect")
+}
